Zaktualizowanie typu parametru jasności lampy w funkcji writeLampBrightness na Float

+++ app/src/main/java/com/example/lpksmartcenter/data/DevicesDataBase.kt
@@ -40,7 +40,7 @@ class DevicesDataBase {
             .addOnFailureListener { onComplete(false) }
     }
 
-    fun writeLampBrightness(key: String, brightness: Int, onComplete: (Boolean) -> Unit = {}) {
+    fun writeLampBrightness(key: String, brightness: Float, onComplete: (Boolean) -> Unit = {}) {
         database.child(key).setValue(brightness.toLong())
             .addOnSuccessListener { onComplete(true) }
             .addOnFailureListener { onComplete(false) }
